14-a2a-protocal/a2a_friend_schechuling/karley_agent_adk/pyproject.toml
[project]
name = "karley-scheduling-agent"
version = "0.1.0"
description = "Karley's scheduling agent for the A2A project."
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk>=0.2.5",
    "google-adk>=1.2.1",
    "python-dotenv",
    "uvicorn",
] 